Added stub classes and their respective JUnit test classes
We implemented some JUnit test cases for TransitionEntity, but not all
of them.

We do NOT need to implement the test cases for the other test classes,
we just need them as stubs for Deliverable 3.

We just need to complete TestStateEntity and TestTransitionEntity.  We
need test cases for phase 2 shit, but do not need to actually implement
the phase two shit, just have stub methods, if any, for them.

Text Sam if you have questions or comments - I'm pretty sure I have a
grasp on this.