#1  Added some Error Checking stuff
ErrorTestingAction now checks that there is exactly one startStateFigure
and one EndStateFigure, and brings up a dialog box if there is a
problem.  Found that start and end state entities were not being added
to the bucket in the clone method, so I fixed this.

There are some TODO's listed in the ErrorTestingAction that outline some
of the errors we will need to check for.  Dont mind the "cheat"
comments, this is just for deliverable 7 purposes.