what the hell is this?
So apparently startStateEntities were not being added to the bucket, but
rather this "start" list?  i dont see any recent changes to application
model so i have no idea how this got here or what it's doing.  the error
checking uses the bucket to check for 1 start and end state so we need
it in there for error checking to function, but revert this if we
actually dont want the start state in the bucket.